requirements.txt
flask
peewee
pyyaml
click
flask_swagger
flask_swagger_ui
